# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnMate AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a document (base64 encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with AI-generated details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}/process:
    post:
      operationId: processDocument
      tags: [documents]
      summary: Process a document with AI (generates summary, flashcards, quiz)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        filename:
          type: string
        summary:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, processed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - filename
        - summary
        - status
        - createdAt
    DocumentWithDetails:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        filename:
          type: string
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        keyConcepts:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, processing, processed, failed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - filename
        - content
        - summary
        - keyConcepts
        - status
        - createdAt
    DocumentInput:
      type: object
      properties:
        title:
          type: string
        filename:
          type: string
        content:
          type: string
          description: Base64-encoded file content or raw text
      required:
        - title
        - filename
        - content
...